placable buildings

+++ Resources/Shader/material.frag
@@ -4,7 +4,7 @@ struct Material {
 	vec3 ambient;
 	vec3 diffuse;
 	vec3 specular;
-	float shininess;
+	int shininess;
 	float opaqueness;
 };
